docs(Table): expand section about Table Column
